# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KPI Command Center",
    "description": "Operaciones ATM, sucursales y planificación geoespacial. Importes en DOP.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/catalogs/provinces": {
      "get": {
        "tags": [
          "Catálogos"
        ],
        "summary": "Provinces",
        "operationId": "provinces_api_v1_catalogs_provinces_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProvinceOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Catálogos"
        ],
        "summary": "Create Province",
        "operationId": "create_province_api_v1_catalogs_provinces_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProvinceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvinceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalogs/filters": {
      "get": {
        "tags": [
          "Catálogos"
        ],
        "summary": "Filters",
        "operationId": "filters_api_v1_catalogs_filters_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/branches": {
      "get": {
        "tags": [
          "Sucursales y colas"
        ],
        "summary": "List Branches",
        "operationId": "list_branches_api_v1_branches_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provincia",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provincia"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BranchStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Estado"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_BranchOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sucursales y colas"
        ],
        "summary": "Create Branch",
        "operationId": "create_branch_api_v1_branches_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/branches/{id}": {
      "get": {
        "tags": [
          "Sucursales y colas"
        ],
        "summary": "Detail",
        "operationId": "detail_api_v1_branches__id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Sucursales y colas"
        ],
        "summary": "Patch Branch",
        "operationId": "patch_branch_api_v1_branches__id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/branches/{id}/atms": {
      "get": {
        "tags": [
          "Sucursales y colas"
        ],
        "summary": "Branch Atms",
        "operationId": "branch_atms_api_v1_branches__id__atms_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ATMOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/branches/{id}/readings": {
      "post": {
        "tags": [
          "Sucursales y colas"
        ],
        "summary": "Ingest Reading",
        "operationId": "ingest_reading_api_v1_branches__id__readings_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchReadingCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchReadingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Sucursales y colas"
        ],
        "summary": "Readings",
        "operationId": "readings_api_v1_branches__id__readings_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_BranchReadingOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/branches/{id}/financials": {
      "post": {
        "tags": [
          "Sucursales y colas"
        ],
        "summary": "Ingest Financial",
        "operationId": "ingest_financial_api_v1_branches__id__financials_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinancialCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancialOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Sucursales y colas"
        ],
        "summary": "Financials",
        "operationId": "financials_api_v1_branches__id__financials_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_FinancialOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BranchCreate": {
        "properties": {
          "codigo": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Codigo"
          },
          "nombre": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Nombre"
          },
          "provincia": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Provincia"
          },
          "municipio": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Municipio"
          },
          "direccion": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Direccion"
          },
          "latitud": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitud"
          },
          "longitud": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitud"
          },
          "estado": {
            "$ref": "#/components/schemas/BranchStatus",
            "default": "OPERATIVA"
          },
          "total_cajeros_humanos": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Total Cajeros Humanos",
            "default": 6
          },
          "sla_objetivo_segundos": {
            "type": "integer",
            "maximum": 86400.0,
            "exclusiveMinimum": 0.0,
            "title": "Sla Objetivo Segundos",
            "default": 600
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "codigo",
          "nombre",
          "provincia",
          "municipio",
          "direccion",
          "latitud",
          "longitud"
        ],
        "title": "BranchCreate"
      },
      "BranchOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "codigo": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Codigo"
          },
          "nombre": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Nombre"
          },
          "provincia": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Provincia"
          },
          "municipio": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Municipio"
          },
          "direccion": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Direccion"
          },
          "latitud": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitud"
          },
          "longitud": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitud"
          },
          "estado": {
            "$ref": "#/components/schemas/BranchStatus",
            "default": "OPERATIVA"
          },
          "total_cajeros_humanos": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Total Cajeros Humanos",
            "default": 6
          },
          "sla_objetivo_segundos": {
            "type": "integer",
            "maximum": 86400.0,
            "exclusiveMinimum": 0.0,
            "title": "Sla Objetivo Segundos",
            "default": 600
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "codigo",
          "nombre",
          "provincia",
          "municipio",
          "direccion",
          "latitud",
          "longitud",
          "created_at",
          "updated_at"
        ],
        "title": "BranchOut"
      },
      "BranchPatch": {
        "properties": {
          "nombre": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre"
          },
          "direccion": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Direccion"
          },
          "estado": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BranchStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "total_cajeros_humanos": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Cajeros Humanos"
          },
          "sla_objetivo_segundos": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 86400.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Objetivo Segundos"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "BranchPatch"
      },
...
      "BranchStatus": {
        "type": "string",
        "enum": [
          "OPERATIVA",
          "MANTENIMIENTO",
          "CERRADA"
        ],
        "title": "BranchStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_ATMOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ATMOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ATMOut]"
      },
...
      "Page_BranchOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BranchOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[BranchOut]"
      },
...
      "Page_ProvinceOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProvinceOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ProvinceOut]"
      },
...
      "ProvinceCreate": {
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Nombre"
          },
          "region": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Region"
          },
          "poblacion": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Poblacion"
          },
          "superficie_km2": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Superficie Km2"
          },
          "fuente": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Fuente"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nombre",
          "region"
        ],
        "title": "ProvinceCreate"
      },
      "ProvinceOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Nombre"
          },
          "region": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Region"
          },
          "poblacion": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Poblacion"
          },
          "superficie_km2": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Superficie Km2"
          },
          "fuente": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Fuente"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "nombre",
          "region"
        ],
        "title": "ProvinceOut"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}